change preference value type to float

+++ by.Reba/Core/DataTransferObjects/Positivity/PositivityDTO.cs
@@ -4,6 +4,6 @@
     {
         public Guid Id { get; set; }
         public string Title { get; set; }
-        public int Value { get; set; }
+        public float Value { get; set; }
     }
 }
